Split api views into multiple files
