+++ scoring_engine/web/templates/service.html
@@ -119,7 +119,7 @@
             function refreshServicesNavbar() {
               $.ajax({
                   cache: false,
-                  url: "/api/team/{{service.team.id}}/services_status",
+                  url: "/api/team/{{service.team.id}}/services/status",
                   dataType: 'json',
                   success: function(data) {
                     navbar_str = ""
@@ -145,6 +145,7 @@
             }
 
             $(document).ready(function() {
+              refreshServicesNavbar();
               // Disable datatables error reporting
               $.fn.dataTable.ext.errMode = 'none';
 
@@ -154,7 +155,7 @@
                 })
                 .DataTable( {
                   "ajax": {
-                    "url": "/api/service/get_checks/{{ id }}",
+                    "url": "/api/service/{{ id }}/checks",
                     "type": 'GET'
                   },
                   "columns": [
@@ -247,8 +248,6 @@
                   table.ajax.reload(expandRows, false);
                   refreshServicesNavbar();
                 }, 30000);
-
-                refreshServicesNavbar();
             });
           </script>
         </div>
